sous-chef/pyproject.toml
[tool.poetry]
name = "sous_chef"
version = "1.2.0"
description = "Plans your meals and grocery list"
authors = ["Aurorian Clouds"]

[tool.poetry.dependencies]
python = "^3.9"
feather-format = "^0.4.1"
fuzzywuzzy = "^0.18.0"
google-api-python-client = "^1.12.8"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.2"
# future version of pandas removes append
pandas = "1.3.5"
pint = "^0.16.1"
pygsheets = "^2.0.5"
python-Levenshtein = "^0.12.2"
pyyaml = "^5.3.1"
todoist-python = "^8.1.3"
hydra-core = "^1.1.1"
structlog = "^21.5.0"
regex = "^2022.1.18"
pandera = "^0.9.0"

[tool.poetry.dev-dependencies]
coverage = "^6.3"
pytest = "^6.2.5"
pytest-structlog = "^0.5"
freezegun = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
